# planejamento/connect_api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EBT Connect - contrato candidato",
    "version": "0.1.0-planejado",
    "description": "Especificação futura. Nenhum endpoint foi implementado. DTO/policy da fonte precisam ser confrontados em P01/P11."
  },
  "x-estado": "planejado",
  "paths": {
    "/api/connect/v1/conversations/{conversationId}/messages": {
      "parameters": [
        {
          "name": "conversationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1
          }
        }
      ],
      "get": {
        "operationId": "listConversationMessages",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página autorizada, cursor estável",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagePage"
                }
              }
            }
          },
          "400": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "queueConversationReply",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Persistida/enfileirada; não significa envio/entrega",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedReply"
                }
              }
            }
          },
          "200": {
            "description": "Repetição da mesma operação já existente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedReply"
                }
              }
            }
          },
          "428": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "400": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/connect/v1/messages/{messageId}": {
      "get": {
        "operationId": "getMessageState",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado observado e evidência disponível",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/connect/meta/{appConnectionKey}": {
      "parameters": [
        {
          "name": "appConnectionKey",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "minLength": 1
          }
        }
      ],
      "get": {
        "operationId": "verifyMetaWebhook",
        "security": [],
        "parameters": [
          {
            "name": "hub.mode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "hub.verify_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "hub.challenge",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Somente desafio validado; sem efeitos de mensagem",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "receiveMetaWebhook",
        "security": [
          {
            "metaSignature": []
          }
        ],
        "description": "Validar assinatura nos bytes originais, mapear conta/conexão e persistir duravelmente antes de ACK. JSON do provedor permanece externo ao domínio.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Envelope recebido duravelmente ou duplicata já durável; não é resposta ao cliente"
          },
          "400": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Erro seguro, código e traceId; sem segredos/conteúdo privado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Credencial candidata da API; estratégia de sessão/provedor deve ser fechada em P04. Não embutir token."
      },
      "metaSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Hub-Signature-256",
        "description": "Representação documental de assinatura HMAC; não é uma API key. Validar bytes originais e segredo de app conforme contrato vigente do provedor."
      }
    },
    "schemas": {
      "ReplyRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "content",
          "contentType"
        ],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1
          },
          "contentType": {
            "type": "string",
            "const": "text"
          },
          "replyToMessageId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "QueuedReply": {
        "type": "object",
        "required": [
          "messageId",
          "operationId",
          "status",
          "statusUrl"
        ],
        "properties": {
          "messageId": {
            "type": "string",
            "minLength": 1
          },
          "operationId": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "accepted",
              "sent",
              "delivered",
              "read",
              "failed",
              "unknown"
            ]
          },
          "statusUrl": {
            "type": "string"
          },
          "version": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Message": {
        "type": "object",
        "required": [
          "id",
          "conversationId",
          "direction",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "conversationId": {
            "type": "string",
            "minLength": 1
          },
          "direction": {
            "type": "string",
            "enum": [
              "incoming",
              "outgoing",
              "internal"
            ]
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "received",
              "queued",
              "accepted",
              "sent",
              "delivered",
              "read",
              "failed",
              "unknown"
            ]
          },
          "providerMessageId": {
            "type": [
              "string",
              "null"
            ]
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MessagePage": {
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "traceId": {
            "type": "string",
            "minLength": 1
          },
          "detail": {
            "type": "string"
          }
        }
      }
    }
  }
}